+++ lib/dnsimple/client/clients.rb
@@ -23,11 +23,6 @@ module Dnsimple
       @services[:domains] ||= Client::DomainsService.new(self)
     end
 
-    # @return [Dnsimple::Client::DomainServicesService] The domain-services-related API proxy.
-    def domain_services
-      @services[:domains] ||= Client::DomainServicesService.new(self)
-    end
-
     # @return [Dnsimple::Client::IdentityService] The identity-related API proxy.
     def identity
       @services[:auth] ||= Client::IdentityService.new(self)